Improve primary navigation accessibility
Using "display: none;" is bad for accessibility because it hides the
element from screen-reading software.

It was also unnecessary in this case. Setting a width and height of zero
and "overflow: hidden;" as the default values for the navbar is enough
to hide while still making it visible to screen-reading software.

Users can now skip to the navigation landmark without having to toggle
the menu beforehand.