Improve accessibility of primary navigation
On smaller screens, the primary navigation menu is now positioned
off-screen when not toggled or focused. Before, hidden menu items were
still selectable when sliding the finger across the screen of a mobile
device with VoiceOver/TalkBack enabled, which might be confusing.

Now, users of AT on mobile devices can either toggle the menu and use
the on-screen menu items or find the off-screen navigation landmark or
menu items by selecting the respective VoiceOver rotor or TalkBack
reading control setting.